moved the cart page to the Pages folder and usercreation to the components
